+++ modules/Assets/assets/dialogs/asset-picker.js
@@ -11,6 +11,10 @@ export default {
     props: {
         filter: {
             default: null
+        },
+        multiple: {
+            type: Boolean,
+            default: false
         }
     },
 
@@ -28,15 +32,15 @@ export default {
             </div>
 
             <div class="kiss-margin">
-                <assets-manager :modal="true" :selectAsset="selectAsset" :initFilter="filter"></assets-manager>
+                <assets-manager :modal="true" :onSelect="onSelect" :selectMultiple="multiple" :initFilter="filter"></assets-manager>
             </div>
 
         </div>
     `,
 
     methods: {
-        selectAsset(asset) {
-            this.$call('selectAsset', asset);
+        onSelect(selected) {
+            this.$call('onSelect', selected);
             this.$close();
         }
     }