+++ modules/System/assets/vue-components/fields-renderer.js
@@ -80,12 +80,18 @@ let FieldRenderer = {
                 this.val = [];
             }
 
-            this.fieldItem = {
-                field,
-                value: JSON.parse(JSON.stringify(field.default || null)),
-                create: true,
-                meta: this.fieldTypes[field.type]
+            let defaultAction = (field) => {
+
+                this.fieldItem = {
+                    field,
+                    value: JSON.parse(JSON.stringify(field.default || null)),
+                    create: true,
+                    meta: this.fieldTypes[field.type]
+                };
             };
+
+            (this.fieldTypes[field.type]?.addFieldItem || defaultAction)(field, this.val);
+
         },
 
         editFieldItem(field, index) {
@@ -158,6 +164,7 @@ let FieldRenderer = {
             <component :is="getFieldType()" v-model="val" v-bind="field.opts" :locale="locale" :data-field-render-uid="uid" v-if="!field.multiple"></component>
 
             <div v-if="field.multiple">
+
                 <kiss-card class="kiss-padding-small kiss-size-small kiss-color-muted" theme="bordered contrast" v-show="!val || !Array.isArray(val) || !val.length">{{ t('No items') }}</kiss-card>
 
                 <vue-draggable v-model="val" handle=".fm-handle" v-if="Array.isArray(val)">
@@ -165,7 +172,7 @@ let FieldRenderer = {
                         <div class="kiss-margin-small kiss-flex kiss-flex-middle">
                             <kiss-card class="kiss-flex-1 kiss-padding-small kiss-size-small kiss-position-relative" theme="bordered contrast">
                                 <span class="kiss-badge kiss-badge-outline kiss-color-muted" v-if="val[index] == null">n/a</span>
-                                <div class="kiss-text-truncate" v-else-if="fieldTypes[field.type] && fieldTypes[field.type].render" v-html="fieldTypes[field.type].render(val[index], field)"></div>
+                                <div class="kiss-text-truncate" v-else-if="fieldTypes[field.type]?.render" v-html="fieldTypes[field.type].render(val[index], field)"></div>
                                 <div v-else>
                                     <span class="kiss-badge kiss-badge-outline" v-if="Array.isArray(val[index])">{{ val[index].length }}</span>
                                     <span class="kiss-badge kiss-badge-outline" v-else-if="typeof(val[index]) === 'object'">Object</span>